requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
click==8.3.0
Deprecated==1.3.1
et_xmlfile==2.0.0
fastapi==0.120.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jsonpath-ng==1.7.0
limits==5.6.0
numpy==2.3.4
openpyxl==3.1.5
orjson==3.11.4
packaging==25.0
pandas==2.3.3
ply==3.11
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wrapt==2.0.0
xmltodict==1.0.2
